Switch to derived serialization
